# docs/swagger.yaml
basePath: /
definitions:
  models.AddressInfo:
    properties:
      country:
        type: string
      district:
        type: string
      full_address:
        type: string
      house:
        type: string
      id:
        type: string
      landmark:
        type: string
      pincode:
        type: string
      post_office:
        type: string
      state:
        type: string
      street:
        type: string
      subdistrict:
        type: string
      type:
        type: string
      vtc:
        description: Village/Town/City
        type: string
    type: object
...
  models.CollaboratorProductResponse:
    properties:
      brand_name:
        type: string
      collaborator_id:
        type: string
      collaborator_name:
        type: string
      created_at:
        type: string
      dosage_instructions:
        type: string
      gst_rate:
        type: number
      hsn_code:
        type: string
      id:
        type: string
      images:
        description: Parsed from JSON
        items:
          type: string
        type: array
      is_active:
        type: boolean
      product:
        allOf:
        - $ref: '#/definitions/models.ProductSummary'
        description: Embedded product info
      product_id:
        type: string
      product_name:
        type: string
      product_sku:
        type: string
      updated_at:
        type: string
      usage_details:
        type: string
    type: object
  models.CollaboratorResponse:
    properties:
      address:
        allOf:
        - $ref: '#/definitions/models.AddressInfo'
        description: Embedded from AAA
      address_id:
        type: string
      bank_account_no:
        type: string
      bank_ifsc:
        type: string
      bank_name:
        type: string
      company_name:
        type: string
      contact_number:
        type: string
      contact_person:
        type: string
      created_at:
        type: string
      email:
        type: string
      experience:
        type: string
      external_id:
        type: string
      gst_number:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      logo:
        type: string
      pan_number:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateCollaboratorProductRequest:
    properties:
      brand_name:
        type: string
      dosage_instructions:
        type: string
      gst_rate:
        maximum: 100
        minimum: 0
        type: number
      hsn_code:
        description: HSN codes are 8 digits
        type: string
      images:
        description: Array of S3 paths
        items:
          type: string
        type: array
      product_id:
        type: string
      usage_details:
        type: string
    required:
    - brand_name
    - gst_rate
    - hsn_code
    - product_id
    type: object
  models.CreateCollaboratorRequest:
    properties:
      address:
        allOf:
        - $ref: '#/definitions/models.CreateAddressRequest'
        description: Inline address creation via AAA
      bank_account_no:
        type: string
      bank_ifsc:
        type: string
      bank_name:
        type: string
      company_name:
        type: string
      contact_number:
        type: string
      contact_person:
        type: string
      email:
        type: string
      experience:
        type: string
      gst_number:
        type: string
      logo:
        type: string
      pan_number:
        type: string
    required:
    - bank_account_no
    - bank_ifsc
    - company_name
    - contact_number
    - contact_person
    type: object
...
  models.CreateProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  models.ProductResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.ProductSummary:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  models.UpdateCollaboratorProductRequest:
    properties:
      brand_name:
        type: string
      dosage_instructions:
        type: string
      gst_rate:
        type: number
      hsn_code:
        type: string
      images:
        items:
          type: string
        type: array
      is_active:
        type: boolean
      usage_details:
        type: string
    type: object
...
  utils.ErrorResponseModel:
    properties:
      error:
        example: Detailed error message
        type: string
      message:
        example: Error occurred
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: info@kisanlink.in
    name: Kisanlink Support
    url: https://github.com/Kisanlink/fpo-erp
  description: Comprehensive ERP system for agricultural cooperatives with multi-tenant
    architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Kisanlink ERP API
  version: "1.0"
paths:
...
  /api/v1/collaborator-products/{id}:
    delete:
      description: Delete collaborator-product association by ID (soft delete, requires
        authentication)
      parameters:
      - description: 'Collaborator Product ID (format: CPRD_xxxxxxxx)'
        example: CPRD_12345678
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Association deleted successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Association not found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Delete Collaborator Product
      tags:
      - Collaborator Products
    get:
      description: Retrieve a specific collaborator-product association
      parameters:
      - description: 'Collaborator Product ID (format: CPRD_xxxxxxxx)'
        example: CPRD_12345678
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Association retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CollaboratorProductResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Association not found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      summary: Get Collaborator Product
      tags:
      - Collaborator Products
    put:
      consumes:
      - application/json
      description: Update collaborator-product metadata (requires authentication)
      parameters:
      - description: 'Collaborator Product ID (format: CPRD_xxxxxxxx)'
        example: CPRD_12345678
        in: path
        name: id
        required: true
        type: string
      - description: Updated metadata
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCollaboratorProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Association updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CollaboratorProductResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Association not found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Update Collaborator Product
      tags:
      - Collaborator Products
  /api/v1/collaborators:
    get:
      description: Retrieve all collaborators (requires authentication)
      produces:
      - application/json
      responses:
        "200":
          description: Collaborators retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CollaboratorResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get All Collaborators
      tags:
      - Collaborators
    post:
      consumes:
      - application/json
      description: Create a new collaborator/vendor (requires authentication)
      parameters:
      - description: Collaborator data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateCollaboratorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Collaborator created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CollaboratorResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Create Collaborator
      tags:
      - Collaborators
...
  /api/v1/collaborators/{id}/products:
    get:
      description: Retrieve all products for a specific collaborator
      parameters:
      - description: 'Collaborator ID (format: CLAB_xxxxxxxx)'
        example: CLAB_12345678
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Products retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CollaboratorProductResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Collaborator not found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      summary: Get Products by Collaborator
      tags:
      - Collaborator Products
    post:
      consumes:
      - application/json
      description: Associate a product with a collaborator (requires authentication)
      parameters:
      - description: 'Collaborator ID (format: CLAB_xxxxxxxx)'
        example: CLAB_12345678
        in: path
        name: id
        required: true
        type: string
      - description: Product association data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateCollaboratorProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product added successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CollaboratorProductResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Add Product to Collaborator
      tags:
      - Collaborator Products
  /api/v1/collaborators/{id}/products/{product_id}:
    delete:
      description: Remove product association from collaborator (soft delete, requires
        authentication)
      parameters:
      - description: 'Collaborator ID (format: CLAB_xxxxxxxx)'
        example: CLAB_12345678
        in: path
        name: id
        required: true
        type: string
      - description: 'Product ID (format: PROD_xxxxxxxx)'
        example: PROD_12345678
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product removed successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Association not found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Remove Product from Collaborator
      tags:
      - Collaborator Products
...
  /api/v1/products:
    get:
      description: Retrieve all products (requires authentication)
      produces:
      - application/json
      responses:
        "200":
          description: Products retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ProductResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get All Products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product (requires authentication)
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Create Product
      tags:
      - Products
...
  /api/v1/products/{id}/collaborators:
    get:
      description: Retrieve all collaborators for a specific product
      parameters:
      - description: 'Product ID (format: PROD_xxxxxxxx)'
        example: PROD_12345678
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Collaborators retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CollaboratorProductResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      summary: Get Collaborators by Product
      tags:
      - Collaborator Products
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"